now displaying loading message while retrieving user data
